modify delete 권한체크

+++ src/main/java/com/sbs/exam/demo/controller/UsrArticleController.java
@@ -2,13 +2,17 @@ package com.sbs.exam.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.sbs.exam.demo.service.ArticleService;
+import com.sbs.exam.demo.util.Ut;
 import com.sbs.exam.demo.vo.Article;
+import com.sbs.exam.demo.vo.ResultData;
 
 @Controller
 public class UsrArticleController {
@@ -17,55 +21,110 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public Article doAdd(String title, String body) {
-		Article article = articleService.writeArticle(title, body);
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
 		
-		return article;
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if(isLogined == false) {
+			return ResultData.from("F-3", "로그인 후 이용해주세요.");
+		}
+		
+		if(Ut.empty(title)) {
+			return ResultData.from("F-1", "제목을 입력해주세요");
+		}
+		
+		if(Ut.empty(body)) {
+			return ResultData.from("F-2", "내용을 입력해주세요");
+		}
+		
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
+		int id = writeArticleRd.getData1();
+		Article article = articleService.getArticle(id);
+		return ResultData.newRd(writeArticleRd, article);
 	}
 	
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
-	public List<Article> getArticles() {
-		return articleService.getArticles();
+	public ResultData<List<Article>> getArticles() {
+		List<Article> articles= articleService.getArticles();
+		return ResultData.from("S-1", "게시물 리스트", articles);
 	}
 	
 	@RequestMapping("/usr/article/getArticle")
 	@ResponseBody
-	public Object getArticle(int id) {
+	public ResultData<Article> getArticle(int id) {
 		Article article = articleService.getArticle(id);
 		
 		if(article == null) {
-			return id + "번 게시물은 없는 게시물 입니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
 		
-		return article;
+		return ResultData.from("S-1", Ut.f("%d번 게시물이 생성되었습니다.", id), article);
 	}
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public String doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if(isLogined == false) {
+			return ResultData.from("F-2", "로그인 후 이용해주세요.");
+		}
+		
+		
 		Article article = articleService.getArticle(id);
 		
+		if(article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-3", "권한이 없습니다.");
+		}
+		
 		if(article == null) {
-			return id + "번 게시물은 없는 게시물입니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
 		
 		articleService.deleteArticle(id);
 		
-		return id + "번 게시물이 삭제되었습니다.";
+		return ResultData.from("S-2", Ut.f("%d번 게시물이 삭제되었습니다.", id),id);
 	}
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public String doModify(int id, String title, String body) {
+	public ResultData<Integer> doModify(HttpSession httpSession, int id, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if(isLogined == false) {
+			return ResultData.from("F-3", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticle(id);
 		
+		if(article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-4", "권한이 없습니다.");
+		}
+		
 		if(article == null) {
-			return id + "번 게시물은 없는 게시물입니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
 		
 		articleService.modifyArticle(id, title, body);
 		
-		return id + "번 게시물이 수정되었습니다." + "\n" + article;
+		return ResultData.from("S-1", Ut.f("%d번 게시물을 수정하였습니다.", id), id);
 	}
 }
